Un par de fixes en la generación de chunks

+++ core/chunkGeneration/ChunkGeneration.h
@@ -12,7 +12,7 @@ namespace ChunkGeneration {
 
 using ChunkArray = std::vector<std::pair<glm::ivec2, ChunkRenderUnit*>>;
 
-ChunkRenderUnit* GenerateChunk(WorldGenerator& fractal, float x_offset, float z_offset);
+ChunkRenderUnit* GenerateChunk(WorldGenerator& fractal, int x_offset, int z_offset);
 void RegenerateChunks(ChunkArray& chunks, WorldGenerator& generator, int render_distance, glm::vec3 camera_pos);
 
 /* Debug */
